[refactor]: 예약 수정 페이지 CSR 구조로 명확화
- /reservation/[reservationId] 페이지를 CSR로 설계 전환
- 최상단 Page에서 'use client' 선언하여 GNB 상태 제어
- SSR 구조 제거, 클라이언트 중심 상태 관리 방식으로 리팩토링 시작
