+++ src/app/reservation/[reservationId]/page.tsx
@@ -1,7 +1,24 @@
-import ReservationEditWrapper from "@/features/reservation/ui/ReservationWrapper";
+"use client";
 
-function ReservationEditRoutePage() {
-  return <ReservationEditWrapper />;
+import { useEffect } from "react";
+import { useGNBStore } from "@/features/common/store/useGnbStore";
+import ReservationEdit from "@/features/reservation/ui/ReservationEdit";
+import { TopBar } from "@/features/common/components/topbar";
+
+function Page() {
+  const setShowGNB = useGNBStore((state) => state.setShowGNB);
+
+  useEffect(() => {
+    setShowGNB(false);
+    return () => setShowGNB(true);
+  }, [setShowGNB]);
+
+  return (
+    <div className="-mx-4 p-4 flex-1">
+      <TopBar showShare={false} showCart={false} message="예약 수정" />
+      <ReservationEdit />
+    </div>
+  );
 }
 
-export default ReservationEditRoutePage;
+export default Page;
