--- src/features/reservation/components/ReservationItem.tsx
@@ -1,5 +1,0 @@
-function ReservationItem() {
-  return <div>123123123</div>;
-}
-
-export default ReservationItem;
