--- src/features/reservation/ui/ReservationWrapper.tsx
@@ -1,24 +1,0 @@
-"use client";
-
-import { useEffect } from "react";
-import { useGNBStore } from "@/features/common/store/useGnbStore";
-import ReservationEdit from "./ReservationEdit";
-import { TopBar } from "@/features/common/components/topbar";
-
-function ReservationEditWrapper() {
-  const setShowGNB = useGNBStore((state) => state.setShowGNB);
-
-  useEffect(() => {
-    setShowGNB(false);
-    return () => setShowGNB(true);
-  }, [setShowGNB]);
-
-  return (
-    <div className="-mx-4 p-4 flex-1">
-      <TopBar showShare={false} showCart={false} message="예약 수정" />
-      <ReservationEdit />
-    </div>
-  );
-}
-
-export default ReservationEditWrapper;
